feat(framework): frameworkのメインとなる関数を持つファイルを作成
