+++ framework/srcs/framework.c
@@ -1,5 +1,15 @@
 #include "libunit.h"
 
+void	load_test(t_unit_test **list, const char *description, t_test_func f)
+{
+	t_unit_test	*test;
+
+	test = test_init(description, f);
+	if (list == NULL || test == NULL)
+		error_exit();
+	test_add_front(list, test);
+}
+
 int	launch_tests(t_unit_test **test)
 {
 	t_unit_test	*target;
